Extract LazyParameter to separate file
